refactor: move String utils into BlockingClassloader
