+++ modules/core/kotlin/com/github/timaxinc/kcipt/classloader/BlockingClassloader.kt
@@ -1,6 +1,33 @@
 package com.github.timaxinc.kcipt.classloader
 
 abstract class BlockingClassloader(protected val softMode: Boolean, parent: ClassLoader?) : ClassLoader(parent) {
+
+    /**
+     * Checks uf there if the passed element starts with any of the members of the MutableList.
+     *
+     * @param list
+     *          the List with the String which shall be used for the check
+     *
+     * @return
+     *          {@code true} if the String starts with at least one of the members of the members of the List; {@code
+     *          false} otherwise
+     */
+    protected infix fun String.startsWithMember(list: List<String>): Block {
+        list.forEach {
+            if (this.startsWith(it)) {
+                if (this == it) return Block.Exact
+                return Block.Package(it)
+            }
+        }
+        return Block.None
+    }
+
+    protected sealed class Block {
+        object None : Block()
+        object Exact : Block()
+        class Package(val packageName: String) : Block()
+    }
+
     /**
      * Abstract representation of an exception thrown, if the blacklist has prevented the requested operation.
      *
